Added support for decoration of the osgTerrain::DataSet generated databases with a CoordinateSystemNode which reflect the coordinate system of the database.
Added support for reading and writing CoordianteSystemNode into the .osg
and .ive formats.

+++ src/osgDB/Field.cpp
@@ -365,7 +365,7 @@ bool Field::getFloat(double& f) const
     getFieldType();
     if (_fieldType==REAL || _fieldType==INTEGER)
     {
-        f = (float)atof(_fieldCache);
+        f = atof(_fieldCache);
         return true;
     }
     else
